Fix Event Bug
